+++ packages/providers/kubernetes/__init__.py
@@ -1,0 +1,3 @@
+"""
+STARCORE Kubernetes Provider
+"""